requirements.txt
streamlit>=1.37.0
playwright>=1.45.0
python-dotenv
chromium
groq>=0.9.0
beautifulsoup4>=4.12.0